refact: Extracted syles, hooks & new component from AssignmentForm
