+++ Web Ui/src/sections/Assignments/components/AssignmentForm.tsx
@@ -1,4 +1,3 @@
-import { useState, useRef, useEffect } from "react";
 import Button from "@mui/material/Button";
 import {
   Dialog, DialogActions, DialogContent, DialogTitle,
@@ -7,97 +6,14 @@ import {
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import Filter from "./DatePicker";
-import { CreateAssignments } from "../../../modules/Assignments/application/CreateAssingment";
-import AssignmentsRepository from "../../../modules/Assignments/repository/AssignmentsRepository";
-import GetGroups from "../../../modules/Groups/application/GetGroups";
-import { GroupDataObject } from "../../../modules/Groups/domain/GroupInterface";
-import GroupsRepository from "../../../modules/Groups/repository/GroupsRepository";
-import { SelectChangeEvent } from "@mui/material/Select";
-import { Warning, CheckCircle } from "@mui/icons-material";
-import { useGlobalState } from "../../../modules/User-Authentication/domain/authStates";
+import ValidationDialog from "./ValidationDialog";
+import { useAssignmentForm } from "../hooks/useAssignmentForm";
+import { useGroups } from "../hooks/useGroups";
+import { styles } from "../styles/Form.styles";
 
-const styles = {
-  validationDialog: {
-    title: {
-      base: {
-        display: "flex",
-        alignItems: "center",
-        gap: 1.5,
-        fontSize: "1rem",
-        fontWeight: 400,
-        py: 2,
-        fontFamily: '"Roboto","Helvetica","Arial",sans-serif',
-      },
-      error: { color: "#d32f2f" },
-      success: { color: "#2e7d32" },
-    },
-    icon: {
-      error: { color: "#d32f2f", fontSize: 22 },
-      success: { color: "#2e7d32", fontSize: 22 },
-    },
-    actions: { pb: 2, pr: 2 },
-    closeButton: {
-      error: { color: "#d32f2f", textTransform: "none" as const, fontSize: "0.875rem" },
-      success: { color: "#2e7d32", textTransform: "none" as const, fontSize: "0.875rem" },
-    },
-  },
-  form: {
-    dialogTitle: { fontSize: "0.8rem" },
-    cancelButton: { color: "#555", textTransform: "none" as const },
-    createButton: { textTransform: "none" as const },
-  },
-};
-
-// Componente ValidationDialog
-interface ValidationDialogProps {
-  open: boolean;
-  title: string;
-  closeText: string;
-  onClose: () => void;
-}
-
-const ValidationDialog = ({
-  open,
-  title,
-  closeText,
-  onClose,
-}: ValidationDialogProps) => {
-  const isError = title.toLowerCase().includes("error");
-
-  return (
-    <Dialog open={open} onClose={onClose} maxWidth="xs" fullWidth>
-      <DialogTitle
-        sx={{
-          ...styles.validationDialog.title.base,
-          ...(isError
-            ? styles.validationDialog.title.error
-            : styles.validationDialog.title.success),
-        }}
-      >
-        {isError ? (
-          <Warning sx={styles.validationDialog.icon.error} />
-        ) : (
-          <CheckCircle sx={styles.validationDialog.icon.success} />
-        )}
-        {title}
-      </DialogTitle>
-      <DialogActions sx={styles.validationDialog.actions}>
-        <Button
-          onClick={onClose}
-          style={
-            isError
-              ? styles.validationDialog.closeButton.error
-              : styles.validationDialog.closeButton.success
-          }
-        >
-          {closeText}
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-};
-
-// Componente Form principal
+/**
+ * Este componente es un formulario para crear una Tarea
+ */
 interface CreateAssignmentPopupProps {
   open: boolean;
   handleClose: () => void;
@@ -105,168 +21,21 @@ interface CreateAssignmentPopupProps {
 }
 
 function Form({ open, handleClose, groupid }: Readonly<CreateAssignmentPopupProps>) {
-  const [save, setSave] = useState(false);
-  const [validationDialogOpen, setValidationDialogOpen] = useState(false);
-  const [validationMessage, setValidationMessage] = useState("Tarea creada exitosamente");
-  const [auth] = useGlobalState("authData");
-  const [assignmentData, setAssignmentData] = useState({
-    id: 0,
-    title: "",
-    description: "",
-    start_date: new Date(),
-    end_date: new Date(),
-    state: "pending",
-    link: "",
-    comment: "",
-    groupid: groupid,
-  });
-  const isCreateButtonClicked = useRef(false);
+  const { groups } = useGroups(open);
 
-  const handleSaveClick = async () => {
-    setSave(true);
-    if (formInvalid()) {
-      return;
-    }
-
-    isCreateButtonClicked.current = true;
-    const assignmentsRepository = new AssignmentsRepository();
-    const createAssignments = new CreateAssignments(assignmentsRepository);
-
-    if (assignmentData.start_date > assignmentData.end_date) {
-      setValidationMessage("Error: La fecha de inicio no puede ser posterior a la fecha de fin");
-      setValidationDialogOpen(true);
-      setSave(false);
-      return;
-    }
-
-    try {
-      const assignments = await assignmentsRepository.getAssignmentsByGroupid(assignmentData.groupid);
-      const duplicateAssignment = assignments.find(
-        (assignment) =>
-          assignment.title.toLowerCase() === assignmentData.title.toLowerCase()
-      );
-
-      if (duplicateAssignment) {
-        setValidationMessage("Error: Ya existe una tarea con el mismo nombre en este grupo");
-        setValidationDialogOpen(true);
-        setSave(false);
-        return;
-      }
-
-      await createAssignments.createAssignment(assignmentData);
-      setValidationMessage("Tarea creada exitosamente");
-      setValidationDialogOpen(true);
-    } catch (error) {
-      if (error instanceof Error) {
-        if (error.message.includes("Limite de caracteres excedido")) {
-          setValidationMessage("Error: El título no puede tener más de 50 caracteres.");
-        } else {
-          setValidationMessage(`Error: ${error.message}`);
-        }
-      } else {
-        setValidationMessage("Error desconocido al crear la tarea.");
-      }
-      setValidationDialogOpen(true);
-    } finally {
-      setSave(false);
-    }
-  };
-
-  const handleUpdateDates = (newStartDate: Date, newEndDate: Date) => {
-    setAssignmentData((prevData) => ({
-      ...prevData,
-      start_date: newStartDate,
-      end_date: newEndDate,
-    }));
-  };
-
-  const handleInputChange = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-    field: string
-  ) => {
-    const { value } = event.target;
-    setAssignmentData((prevData) => ({
-      ...prevData,
-      [field]: value,
-    }));
-  };
-
-  const handleGroupChange = (event: SelectChangeEvent<number>) => {
-    const groupid = event.target.value as number;
-    setAssignmentData((prevData) => ({
-      ...prevData,
-      groupid,
-    }));
-  };
-
-  const handleCancel = () => {
-    handleClose();
-  };
-
-  const formInvalid = () => {
-    return assignmentData.title.trim() === "" || assignmentData.groupid === 0;
-  };
-
-  useEffect(() => {
-    const effectiveGroupId =
-      groupid || Number(localStorage.getItem("selectedGroup") ?? 0) || 0;
-
-    setSave(false);
-    setAssignmentData({
-      id: 0,
-      title: "",
-      description: "",
-      start_date: new Date(),
-      end_date: new Date(),
-      state: "pending",
-      link: "",
-      comment: "",
-      groupid: effectiveGroupId,
-    });
-  }, [open, groupid]);
-
-  const groupRepository = new GroupsRepository();
-  const [groups, setGroups] = useState<GroupDataObject[]>([]);
-
-  useEffect(() => {
-    const fetchGroups = async () => {
-      const getGroups = new GetGroups(groupRepository);
-      let list: GroupDataObject[] = [];
-
-      if (auth?.userRole === "teacher") {
-        const ids = await getGroups.getGroupsByUserId(auth.userid ?? -1);
-        list = (
-          await Promise.all(ids.map((id: number) => getGroups.getGroupById(id)))
-        ).filter(Boolean) as GroupDataObject[];
-      } else if (auth?.userRole === "admin") {
-        list = await getGroups.getGroups();
-      } else if (auth?.userRole === "student") {
-        let ids: number[] = [];
-        try {
-          const fromLS = JSON.parse(localStorage.getItem("userGroups") ?? "[]");
-          if (Array.isArray(fromLS) && fromLS.length) ids = fromLS;
-        } catch {}
-        if (!ids.length) {
-          ids = await getGroups.getGroupsByUserId(auth.userid ?? -1);
-        }
-        list = (
-          await Promise.all(ids.map((id: number) => getGroups.getGroupById(id)))
-        ).filter(Boolean) as GroupDataObject[];
-      } else {
-        list = [];
-      }
-
-      setGroups(list);
-
-      setAssignmentData((prev) => {
-        const keepCurrent = list.some((g) => g.id === prev.groupid);
-        const fallbackId = list[0]?.id ?? 0;
-        return { ...prev, groupid: keepCurrent ? prev.groupid : fallbackId };
-      });
-    };
-
-    if (open) fetchGroups();
-  }, [open, auth?.userRole, auth?.userid]);
+  const {
+    save,
+    assignmentData,
+    validationDialogOpen,
+    validationMessage,
+    formInvalid,
+    handleInputChange,
+    handleGroupChange,
+    handleUpdateDates,
+    handleCancel,
+    handleSaveClick,
+    handleValidationClose,
+  } = useAssignmentForm({ open, groupid, groups, handleClose });
 
   return (
     <Dialog open={open} onClose={handleClose} maxWidth="sm" fullWidth>
@@ -296,11 +65,7 @@ function Form({ open, handleClose, groupid }: Readonly<CreateAssignmentPopupProp
 
             <TextField
               error={save && !assignmentData.title.trim()}
-              helperText={
-                save && !assignmentData.title.trim()
-                  ? "El título es requerido"
-                  : ""
-              }
+              helperText={save && !assignmentData.title.trim() ? "El título es requerido" : ""}
               autoFocus
               margin="dense"
               id="assignment-title"
@@ -355,13 +120,7 @@ function Form({ open, handleClose, groupid }: Readonly<CreateAssignmentPopupProp
           open={validationDialogOpen}
           title={validationMessage}
           closeText="Cerrar"
-          onClose={() => {
-            if (!validationMessage.includes("Error")) {
-              window.location.reload();
-            } else {
-              setValidationDialogOpen(false);
-            }
-          }}
+          onClose={handleValidationClose}
         />
       )}
     </Dialog>
